Create shannon_entropy
Hi there! the following function calculates the Shannon entropy of a given signal. Shannon entropy is a measure of the uncertainty or randomness in the data and is widely used in information theory, signal processing and, why not, neuroscience! This implementation uses the Freedman-Diaconis rule to determine an appropriate number of bins for discretizing the signal, providing a reliable estimate of entropy for a wide range of data distributions.